requirements.txt
fastapi==0.110.0
passlib==1.7.4
pydantic[email]==2.6.4
pydantic_settings==2.2.1
python_jose==3.3.0
SQLAlchemy==2.0.25
SQLAlchemy_Utils==0.41.1
uvicorn==0.28.0
gunicorn
psycopg[binary]
python-multipart
regex==2024.7.24
bcrypt==4.0.1
alembic==1.13.2
pytest==8.3.2
httpx
pytest-asyncio
pytest-dotenv
pytest-cov
logging-extension==0.4.0
